requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
binaryornot==0.4.4
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.4
cssbeautifier==1.15.4
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.3
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.1
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.1
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250822
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
